Split constraint tests into correct paths
