+++ tests/unit/scheduler/constraints/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for scheduler constraints."""
